cmdb/requirements/common.txt
alembic==0.8.10
amqp==2.1.4
aniso8601==1.2.0
appdirs==1.4.0
appnope==0.1.0
billiard==3.5.0.2
blinker==1.4
celery==4.0.2
click==6.7
decorator==4.0.11
Flask==0.12
Flask-Admin==1.4.2
Flask-HTTPAuth==3.2.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.3
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Uploads==0.2.1
funcy==1.7.2
gnureadline==6.3.3
httpie==0.9.9
IPy==0.83
ipython==4.0.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.9.5
kombu==4.0.2
Mako==1.0.6
MarkupSafe==0.23
olefile==0.44
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.13
ptyprocess==0.5.1
Pygments==2.2.0
PyMySQL==0.7.9
pyparsing==2.1.10
python-dateutil==2.6.0
python-editor==1.0.3
pytz==2016.10
redis==2.10.5
requests==2.13.0
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.5
traitlets==4.3.1
urllib3==1.20
vine==1.1.3
wcwidth==0.1.7
Werkzeug==0.11.15
WTForms==2.1
xlrd==1.0.0
